renamed to tritongrader
