+++ src/tritongrader/test_cases.py
@@ -3,7 +3,7 @@ import traceback
 import binascii
 import logging
 
-from grader.utils import *
+from tritongrader.utils import *
 
 
 class TestCaseResult:
