+++ test/test_autograder.py
@@ -1,4 +1,4 @@
-from src.grader.autograder import Autograder
+from src.tritongrader.autograder import Autograder
 
 if __name__ == "__main__":
     ag = Autograder(name="Test Autograder")
